Moderator system

+++ src/main/java/com/blog_java/application/services/CommentService.java
@@ -2,14 +2,20 @@ package com.blog_java.application.services;
 
 import com.blog_java.domain.dtos.comment.CommentRegisterDto;
 import com.blog_java.domain.dtos.comment.UpdateCommentDto;
+import com.blog_java.domain.enums.UserRole;
 import com.blog_java.domain.models.Comment;
 import com.blog_java.domain.models.Post;
+import com.blog_java.domain.models.User;
 import com.blog_java.infra.repositories.CommentRepository;
 import com.blog_java.infra.repositories.PostRepository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.access.AccessDeniedException;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+
+import java.util.Set;
 
 
 @Service
@@ -30,14 +36,16 @@ public class CommentService {
     @Transactional
     public Comment createComment(CommentRegisterDto commentRegisterDto)
     {
-        if(commentRegisterDto.postId().equals("") || commentRegisterDto.comment().equals(""))
+        if(commentRegisterDto.postId() == null || commentRegisterDto.comment().equals(""))
         {
             throw new IllegalArgumentException();
         }
 
         Post post = postService.findPostById(commentRegisterDto.postId());
 
-        Comment comment = new Comment(commentRegisterDto,post);
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        Comment comment = new Comment(commentRegisterDto,post,user);
 
         post.addComment(comment);
 
@@ -61,19 +69,39 @@ public class CommentService {
     }
 
     @Transactional
-    public Comment updateCommentById(Long id, UpdateCommentDto updateCommentDto)
-    {
-        Comment comment =  findCommentById(id);
+    public Comment updateCommentById(Long id, UpdateCommentDto updateCommentDto) {
+        Comment comment = findCommentById(id);
+
+        // Usuário autenticado
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        boolean isOwner = comment.getUser().getId().equals(user.getId());
+        boolean isAdmin = user.getRole().equals(UserRole.ADMIN);
+        boolean isModerator = user.getRole().equals(UserRole.MODERATOR);
+
+
+        if (!isOwner && !isAdmin && !isModerator) {
+            throw new AccessDeniedException("Você não tem permissão para atualizar este comentário.");
+        }
 
         comment.setComment(updateCommentDto.comment());
 
-        return commentRepository. save(comment);
+        return commentRepository.save(comment);
     }
-
     @Transactional
-    public void deleteComment(Long id)
-    {
+    public void deleteComment(Long id) {
         Comment comment = findCommentById(id);
+
+        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        // Permissões em um set
+        Set<UserRole> roles = Set.of(UserRole.ADMIN, UserRole.MODERATOR);
+
+        if (!comment.getUser().getId().equals(user.getId()) &&
+                !comment.getPost().getUser().getId().equals(user.getId()) &&
+                roles.stream().noneMatch(role -> role.equals(user.getRole()))) {
+            throw new AccessDeniedException("Você não tem permissão para excluir este comentário.");
+        }
 
         commentRepository.delete(comment);
     }
